moved vector 2D and MovementCurves
