+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/libraries/MovementCurves/MovementCurvesTest.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.libraries;
+package org.firstinspires.ftc.teamcode.libraries.MovementCurves;
 
 public class MovementCurvesTest {
     public static void main(String[] args) {
